Tokenizer

+++ lib/collections/hash_map.h
@@ -425,4 +425,4 @@ typedef enum HashMapSlotOccupancy {
     IMPL_HASH_MAP_ENTRY_INSERT(KEY, VALUE)                                      \
     IMPL_HASH_MAP_ENTRY_REMOVE(KEY, VALUE)                                      \
 
-DECL_HASH_MAP(u32, i32)
+DECL_HASH_MAP(usize, usize)
